Struct marc::Record

pub struct Record<'a> { /* fields omitted */ }

Parsed MARC Record.

It could be borrowed if it was parsed from a buffer or it could be owned if it was read from an io::Read implementor.

Methods

impl<'a> Record<'a>

fn parse<'x>(input: &'x [u8]) -> Result<Record<'x>>

Will try to parse record from a buffer.

Will borrow an input for the lifetime of a produced record.

fn from_vec<I>(input: I) -> Result<Record<'static>> where I: Into<Vec<u8>>

Will crate owned record from vector of bytes.

fn read<T: Read>(input: &mut T) -> Result<Option<Record<'static>>>

Will try to read a Record from an io::Read implementor.

Will produce owned version of Record.

fn field<T: Into<Tag>>(&self, tag: T) -> Vec<Field>

Will return fields with tag == Tag

fn fields<'r>(&'r self) -> Fields<'r>

Will return iterator over fields of a record

fn as_ref(&self) -> &[u8]

View into a data of a record.

fn record_status(&self) -> RecordStatus

fn type_of_record(&self) -> TypeOfRecord

fn bibliographic_level(&self) -> BibliographicLevel

fn type_of_control(&self) -> TypeOfControl

fn character_coding_scheme(&self) -> CharacterCodingScheme

fn encoding_level(&self) -> EncodingLevel

fn descriptive_cataloging_form(&self) -> DescriptiveCatalogingForm

fn multipart_resource_record_level(&self) -> MultipartResourceRecordLevel