Struct marc::RecordBuilder

pub struct RecordBuilder { /* fields omitted */ }

Record builder.

Methods

impl RecordBuilder

fn new() -> RecordBuilder

Creates default record builer

fn from_record(record: &Record) -> RecordBuilder

Creates record builder from existing record

fn iter_fields(&self) -> Iter<FieldRepr>

Iterator over fields of this builder.

fn add_field<T: Into<FieldRepr>>(&mut self, f: T) -> Result<&mut Self>

A way to add field to this builder.

Errors

Will return error if field is larger than 9.999 bytes.

fn add_fields<T: Into<FieldRepr>>(&mut self, fs: Vec<T>) -> Result<&mut Self>

A way to add multiple fileds to this builder.

Errors

Will return error if any of fields is larger than 9.999 bytes.

fn filter_fields<F>(&mut self, fun: F) -> &mut RecordBuilder where F: FnMut(&Field) -> bool

Will filter fields of this builder by fun predicate.

fn filter_subfields<F>(&mut self, fun: F) -> &mut Self where F: FnMut(&Field, &Subfield) -> bool

Will filter subfields of this builder by fun predicate.

fn get_record(&self) -> Result<Record<'static>>

Returns record.

Errors

Will return error if record length is greater than 99.999 bytes.

fn get_record_status(&self) -> RecordStatus

fn set_record_status(&mut self, x: RecordStatus) -> &mut Self

fn get_type_of_record(&self) -> TypeOfRecord

fn set_type_of_record(&mut self, x: TypeOfRecord) -> &mut Self

fn get_bibliographic_level(&self) -> BibliographicLevel

fn set_bibliographic_level(&mut self, x: BibliographicLevel) -> &mut Self

fn get_type_of_control(&self) -> TypeOfControl

fn set_type_of_control(&mut self, x: TypeOfControl) -> &mut Self

fn get_character_coding_scheme(&self) -> CharacterCodingScheme

fn set_character_coding_scheme(&mut self, x: CharacterCodingScheme) -> &mut Self

fn get_encoding_level(&self) -> EncodingLevel

fn set_encoding_level(&mut self, x: EncodingLevel) -> &mut Self

fn get_descriptive_cataloging_form(&self) -> DescriptiveCatalogingForm

fn set_descriptive_cataloging_form(&mut self, x: DescriptiveCatalogingForm) -> &mut Self

fn get_multipart_resource_record_level(&self) -> MultipartResourceRecordLevel

fn set_multipart_resource_record_level(&mut self, x: MultipartResourceRecordLevel) -> &mut Self